Created a kernel source for testing and extended the bootloader
*Moved all bootloader related files into the "boot" directory
*Created new directory "kernel" that should contain all the files related
directly to the kernel
*Added bootloader module to handle basic reading from a disk device
in real mode using the 0x13 BIOS routine (in boot/disk_io directory)
*Created basic kernel source that should draw one letter ('K') on the
screen
*Extended the bootloader to try to read the sectors after its postion
on the boot device and load them into memory. This is where the kernel
will be 'glued' onto the bootloader binary. This will provide a self-containing
booting image with the kernel. After this it calls the main kernel entry symbol
*Updated the Makefile