Added basic VGA Text Mode driver
*Created new directory for drivers - 'drivers'
*Updated the Makefile with a new targer - 'drivers'
*Beginning of implementation of a basic VGA text mode
driver. For now it can print a char at any position of the screen.
*The driver needs a lot of work to be usable, only two functions are
implemented as of now (printing a single character), and they still need
some more work.
*The Makefile should be rewritten for the most part, it starts to
be unreadable. Wildcard and regular expressions are needed for this.
*The kernel now writes some X's on the screen to test the driver.
*In the drivers I opted to use a structure to hold the attribute of
the character and an enum to hold the colors. This in my opinion
makes the code much more readable and structured, even if it adds
a little overhead/more lines of code.