Added proper CPU exception ISRs.
Kernel:
*Loads the GDT and IDT along with all the exception gates.

IDT:
*Remapped the PIC and enabled interrupts but this causes a CPU exception
(General Protection Exception)

ISR:
*This module will contain all the code for ISR's (exceptions and
hardware interrupts). The isr source and header contain two functions
that should create all the ISRs for exceptions and interrupts.
As of now only the exception creation function is complete.
*As of now the ISR's for CPU exceptions (in the isr/exceptions directory)
are finished. They only dispaly witch exception was raised for
some debugging information.
*The exception ISR functions should also display the error code if it
is pushed onto the stack (needs a int->char* functions for this).
*The hardware interrupt code needs to be written.
*The whole module needs comments.

Bootloader:
*Increased the count of sectors read from 10 to 20 as the kernel binary
grew in size.

IDT&GDT:
*The size passed onto the stack of the IDT and GDT is the actual size in bytes.
The documentation states that it should be (size_in_bytes - 1). I will
need to verify if this doesn't cause the problems with interrupt handling.