Hardware interrupt handling now works.
ISR:
*Added missing hardware interrupt (IRQ) handling code in the
'interrupts/' directory. Update the isr.c function that creates
the IRQ ISRs.
*The interrupt handling code is very simple (only notifies that it is called
by printing text on the screen). They keyboard interrupt needs to be implemented
as the highest priority.
*The interrupt and exception handling function code should be refactored as there is a
lot of the same code over and over again (the Assembly to save the state of the CPU
when a handler is called specified in the ISR).

Kernel:
*Creates the IRQ ISRs after creating the CPU exception ISRs.

IDT:
*Added the missing interrupt enum entries along with their IDT indexes.
*Fixed the passed size of the IDT (-1)

GDT:
*Fixed the passed size of the GDT (-1)