Implemented a basic driver management system.
Driver:
*Added driver.c/driver.h source and header with the driver management
code and library.
*As for now it can register a driver in a table storing its name, type
and address of its buffer (to which it will write or from which it
will read, depending on its type).

Driver/Keyboard:
*Added a public function used for intializing the keyboard driver.
As for now it only contains some code that will be used for testing the
driver management module. This should be expanded later on.

Util/String:
*Created a stub of the 'strcpy' function that is used by the driver
management for copying the drivers name. It only returns NULL for now.
Needs to be fully implemented.

Makefile:
*Expaned the DRIVER_SOURCES and DRIVER_HEADERS expression to contain also
the driver.c/driver.h files.